Consolidate Monday matched adoptee query
